Fix physalloc wraparound code.
physalloc_t::total_size is the number of pages, i.e. bits in
the physalloc_t::map array. physalloc_t::alloc_last_pos is the
index in that array, so the conversion is needed when these two
are compared.

This fixes pagefaults during output to the screen when the page
at va 0xa0000 appears to be inaccessible.