disk_q: add flag_chained to pager_io_request
This flag shows whether the io_request is queued in a disk driver or not.
It must be modified together with 'chain' field under the chain_lock.

flag_pagein and flags_pageout were used previously for this purpose, but
they belong to pager and now pager doesn't queue io_requests anymore.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>